fix: name is not append dev on manifest
