+++ webpack/webpack.common.js
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 
 const path = require('path')
-const CopyPlugin = require('copy-webpack-plugin')
 const SRC_DIR = path.join(__dirname, '..', 'src')
 
 module.exports = {
@@ -35,33 +34,4 @@ module.exports = {
   resolve: {
     extensions: ['.ts', '.tsx', '.js'],
   },
-  plugins: [
-    new CopyPlugin({
-      patterns: [
-        {
-          from: '.',
-          to: '../',
-          context: 'public',
-          transform: (content, path) => {
-            if (!path.includes('manifest')) {
-              return content
-            }
-
-            const manifest = JSON.parse(content.toString())
-            const packageJson = require('../package.json')
-
-            if (this.mode === 'development') {
-              manifest.name += ' (dev)'
-            }
-
-            manifest.version = packageJson.version
-            manifest.description = packageJson.description
-
-            manifest_JSON = JSON.stringify(manifest, null, 2)
-            return manifest_JSON
-          },
-        },
-      ],
-    }),
-  ],
 }
